# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RouteRequest {
  sourceLat: number;
  sourceLng: number;
  destLat: number;
  destLng: number;
  sourceName?: string;
  destName?: string;
}

export interface RoutePoint {
  lat: number;
  lng: number;
}

export type RouteType = (typeof RouteType)[keyof typeof RouteType];

export const RouteType = {
  fastest: "fastest",
  safest: "safest",
} as const;

export type RouteRiskLevel =
  (typeof RouteRiskLevel)[keyof typeof RouteRiskLevel];

export const RouteRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Route {
  type: RouteType;
  points: RoutePoint[];
  safetyScore: number;
  distanceKm: number;
  estimatedMinutes: number;
  riskLevel: RouteRiskLevel;
  color: string;
}

export type RouteResponseTimeOfDayRisk =
  (typeof RouteResponseTimeOfDayRisk)[keyof typeof RouteResponseTimeOfDayRisk];

export const RouteResponseTimeOfDayRisk = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface RouteResponse {
  fastest: Route;
  safest: Route;
  timeOfDayRisk: RouteResponseTimeOfDayRisk;
}

export type RiskScoreResponseLevel =
  (typeof RiskScoreResponseLevel)[keyof typeof RiskScoreResponseLevel];

export const RiskScoreResponseLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type RiskScoreResponseFactors = {
  timeOfDay: number;
  crimeIndex: number;
  lighting: number;
  crowdDensity: number;
};

export interface RiskScoreResponse {
  lat: number;
  lng: number;
  score: number;
  level: RiskScoreResponseLevel;
  factors: RiskScoreResponseFactors;
}

export interface SosRequest {
  lat: number;
  lng: number;
  userId?: string;
  message?: string;
}

export interface SosResponse {
  success: boolean;
  alertId: string;
  message: string;
  contactsNotified: number;
  timestamp: string;
}

export type UnsafeZoneRiskLevel =
  (typeof UnsafeZoneRiskLevel)[keyof typeof UnsafeZoneRiskLevel];

export const UnsafeZoneRiskLevel = {
  medium: "medium",
  high: "high",
} as const;

export interface UnsafeZone {
  id: string;
  name: string;
  lat: number;
  lng: number;
  radius: number;
  riskLevel: UnsafeZoneRiskLevel;
  description: string;
}

export interface UnsafeZonesResponse {
  zones: UnsafeZone[];
}

export type GetRiskScoreParams = {
  lat: number;
  lng: number;
};