# lib/api-zod/src/generated/types/riskScoreResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScoreResponseFactors } from "./riskScoreResponseFactors";
import type { RiskScoreResponseLevel } from "./riskScoreResponseLevel";

export interface RiskScoreResponse {
  lat: number;
  lng: number;
  score: number;
  level: RiskScoreResponseLevel;
  factors: RiskScoreResponseFactors;
}