# lib/api-zod/src/generated/types/route.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */
import type { RoutePoint } from "./routePoint";
import type { RouteRiskLevel } from "./routeRiskLevel";
import type { RouteType } from "./routeType";

export interface Route {
  type: RouteType;
  points: RoutePoint[];
  safetyScore: number;
  distanceKm: number;
  estimatedMinutes: number;
  riskLevel: RouteRiskLevel;
  color: string;
}