# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteRequest {
  sourceLat: number;
  sourceLng: number;
  destLat: number;
  destLng: number;
  sourceName?: string;
  destName?: string;
}