# lib/api-zod/src/generated/types/routeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */
import type { Route } from "./route";
import type { RouteResponseTimeOfDayRisk } from "./routeResponseTimeOfDayRisk";

export interface RouteResponse {
  fastest: Route;
  safest: Route;
  timeOfDayRisk: RouteResponseTimeOfDayRisk;
}