# lib/api-zod/src/generated/types/sosRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */

export interface SosRequest {
  lat: number;
  lng: number;
  userId?: string;
  message?: string;
}