# lib/api-zod/src/generated/types/sosResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */

export interface SosResponse {
  success: boolean;
  alertId: string;
  message: string;
  contactsNotified: number;
  timestamp: string;
}