# lib/api-zod/src/generated/types/unsafeZone.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Women Safety Smart Route API
 * OpenAPI spec version: 0.1.0
 */
import type { UnsafeZoneRiskLevel } from "./unsafeZoneRiskLevel";

export interface UnsafeZone {
  id: string;
  name: string;
  lat: number;
  lng: number;
  radius: number;
  riskLevel: UnsafeZoneRiskLevel;
  description: string;
}